pyproject.toml
[build-system]
requires = ["flit"]
build-backend = "flit_core.buildapi"

[project]
name = "drf-service-layer"
authors = [
    { name = "qu3vipon", email = "qu3vipon@gmail.com" },
]
readme = "README.md"
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Topic :: Internet",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development",
    "Typing :: Typed",
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Web Environment",
    "License :: OSI Approved :: MIT License",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Internet :: WWW/HTTP",
]
requires-python = ">=3.7"
keywords = ["django", "drf", "service", "layer", "service-layer"]
dependencies = [
    'django',
    'djangorestframework',
]
dynamic = ['version', 'description']

[project.urls]
Documentation = "https://github.com/qu3vipon/drf-service-layer"
Source = "https://github.com/qu3vipon/drf-service-layer"

[project.optional-dependencies]
dev = [
    "pytest >=6.2.4,<7.0.0",
    "pytest-django ==4.4.0",
    "pytest-cov >=2.12.0,<3.0.0",
    "flake8 >=3.8.3,<4.0.0",
    "black ==20.8b1",
    "isort >=5.0.6,<6.0.0",
]

[tool.flit.module]
name = "drf_service_layer"